+++ src/utils/render/image.ts
@@ -95,7 +95,7 @@ export async function renderImage(
       x: box.x + s.padding,
       y: box.y + box.height / 2 - s.fontSize / 2,
       size: s.fontSize,
-      font: ctx.latinFont,
+      font: ctx.chineseFont ?? ctx.latinFont,
       color: s.textColor,
     })
   }